adds insert word links script with examples
this script is used to insert a-tags for words. So far used in the kanji
templates for listing links to words using the kanji in question.

links are seperated with a span containing a space.
please not that the seperator is a closure. I was force to do this due
to elements being stored by reference, hence new instances of the span
needs to be inserted into the tags array.