reshapes the usage of adjacent scripts
before the script referred to in anki was the dom printer, now
dom-writers have that function instead.

This removes use of specific ID, the writer appends the markup to the
first article tag.

hence there is a requirement to use article tags in the anki template
when outputting adjacent scripts