adds insertion of word definition data
the insertion of word defintion data was handled in CSS before.
(before I knew about JavaScript being reliable in Anki)

words are stored in tsv files (data/words dir)
the data is stored in files split on the first letter of the word

the tsv data is compiled a JavaScript file with a map ./_words-data.js
(ignored as it is compiled)

the building of the dfn tags is now being done in JavaScript by setting
attributes with data. CSS outputs the tag showing appropriate data based
on the surrounding id (usually a writer e.g. definitions-writer id)