updates example for replacing c[\d] clozes
I updated this so I show an example of a replaced clozed word being in
two places on the markup.

If a word is c[\d] clozed and the word appears later in the template.
The following word will be clozed as well. So no need to use multiple
c[\d] clozes for the same word