adds matching words to base kanji data
for the kanji templates, words using that kanji will be inserted
automatically.

the write-base-kanji-data is now a module that does this.
NOTE: due to write-base-kanji-data being a module, requiring the module
and calling `#main` is needed to get the insertion

I had to use a module here, due to the async nature of a single async
`require([])` call. the data would not be inserted correctly